requirements.txt
torch
torchvision
numpy
tqdm
diffusers
timm
lmdb
Pillow
-e git+https://github.com/LTH14/torch-fidelity.git@master#egg=torch-fidelity
accelerate
swanlab
